pyproject.toml
[project]
name = "arknights_crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "google-cloud-storage>=3.1.0",
    "lxml>=5.3.1",
    "openai>=1.65.2",
]
